extracts reducers for different search results
