+++ client/src/components/OfficialDictionaryResults/OfficialDictionaryResults.js
@@ -1,19 +1,19 @@
 import React from 'react';
+
 export default function OfficialDictionaryResults(props) {
   const { conditionalRender, data, requestStatus } = props;
   const { definitions, etymologies } = data;
+
   return (
     <React.Fragment>
       <div className={ `Result-Section` }>
         <div className={ `Result` }>
           { conditionalRender({
             status: requestStatus,
-            result: (definitions ?
+            result: (definitions &&
               definitions.map((definition, index) => <span
                 key={ `definition-${index}` }>
                 { definition }</span>)
-                :
-                null
             )
           }) }
         </div>
@@ -22,9 +22,9 @@ export default function OfficialDictionaryResults(props) {
         <div className={ `Result` }>
           { conditionalRender({
             status: requestStatus,
-            result: etymologies ?
-              etymologies.map((etymology, index) => <span
-                key={ `etymology-${index}` }>{ etymology }</span>) : null
+            result: etymologies &&
+            etymologies.map((etymology, index) => <span
+              key={ `etymology-${index}` }>{ etymology }</span>)
           }) }
         </div>
       </div>
