+++ client/src/reducers/index.js
@@ -1,8 +1,14 @@
 import { combineReducers } from "redux";
-import dictionarySearch from "./dictionarySearch";
+import searchWord from "./searchWord";
+import officialDictionarySearch from "./officialDictionarySearch";
+import thesaurusTermsSearch from "./thesaurusTermsSearch";
+import urbanDictionarySearch from "./urbanDictionarySearch";
 
 const rootReducer = combineReducers({
-  dictionarySearch
+  searchWord,
+  officialDictionarySearch,
+  thesaurusTermsSearch,
+  urbanDictionarySearch
 });
 
 export default rootReducer;
